fix dtNavMeshQuery crashes

+++ dep/recastnavigation/Detour/Include/DetourNavMeshQuery.h
@@ -160,7 +160,7 @@ public:
 	///  @param[in]		nav			Pointer to the dtNavMesh object to use for all queries.
 	///  @param[in]		maxNodes	Maximum number of search nodes. [Limits: 0 < value <= 65536]
 	/// @returns The status flags for the query.
-	dtStatus init(const dtNavMesh* nav, const int maxNodes, unsigned int threadId);
+	dtStatus init(const dtNavMesh* nav, const int maxNodes, size_t threadId);
 	
 	/// @name Standard Pathfinding Functions
 	// /@{
@@ -523,7 +523,7 @@ private:
 	class dtNodeQueue* m_openList;		///< Pointer to open list queue.
 
 public:
-    unsigned int m_owningThread;
+	size_t m_owningThread;
 };
 
 /// Allocates a query object using the Detour allocator.
