+++ dep/recastnavigation/Detour/Source/DetourNavMeshQuery.cpp
@@ -163,7 +163,7 @@ dtNavMeshQuery::~dtNavMeshQuery()
 /// functions are used.
 ///
 /// This function can be used multiple times.
-dtStatus dtNavMeshQuery::init(const dtNavMesh* nav, const int maxNodes, unsigned int threadId)
+dtStatus dtNavMeshQuery::init(const dtNavMesh* nav, const int maxNodes, size_t threadId)
 {
 	m_nav = nav;
     m_owningThread = threadId;
